requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
pandas==2.1.3
opencv-python==4.8.1.78
deepface==0.0.88
tensorflow==2.16.1
qrcode[pil]==7.4.2
pyzbar==0.1.9
pillow==10.1.0
python-multipart==0.0.6
tf-keras==2.16.0
groq==0.4.2
openpyxl==3.1.2
python-dotenv